refactor(tui): extract status bar and scrollbar component
